ai/requirements.txt
ddgs
llama-index
llama-index-llms-ollama
python-dotenv
requests
xai-sdk
python-dotenv